# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ArtistHub Social Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/conversations/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message (or start conversation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/book/{artistId}:
    post:
      operationId: sendBookingInquiry
      tags: [messages]
      summary: Send a booking inquiry to an artist
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInquiryRequest"
      responses:
        "201":
          description: Inquiry sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /groups:
    get:
      operationId: getGroups
      tags: [groups]
      summary: List groups
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
  /groups/{groupId}/posts:
    get:
      operationId: getGroupPosts
      tags: [groups]
      summary: Get posts for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Group posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"
  /groups/{groupId}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /groups/{groupId}/leave:
    post:
      operationId: leaveGroup
      tags: [groups]
      summary: Leave group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /messages/inquiries/{recipientId}:
    post:
      operationId: sendInquiry
      tags: [messages]
      summary: Send creator inquiry
      parameters:
        - name: recipientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryRequest"
      responses:
        "201":
          description: Inquiry sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    LinkItem:
      type: object
      properties:
        label:
          type: string
        url:
          type: string
      required: [label, url]
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        artistDisplayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        profileType:
          type: string
          enum: [user, artist]
        followerCount:
          type: integer
        friendCount:
          type: integer
        bannerUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        work:
          type: string
          nullable: true
        school:
          type: string
          nullable: true
        about:
          type: string
          nullable: true
        interests:
          type: array
          items:
            type: string
        hasArtistPage:
          type: boolean
        accentColor:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        onboardingStep:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/LinkItem"
        category:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        hasBlockedUser:
          type: boolean
        isBlockedByUser:
          type: boolean
      required: [id, username, profileType, followerCount, friendCount, hasBlockedUser, isBlockedByUser]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        actorSurface:
          type: string
          enum: [personal, artist]
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [public, friends, private]
        repostOfPostId:
          type: integer
          nullable: true
        likeCount:
          type: integer
        isLiked:
          type: boolean
        reactionCounts:
          $ref: "#/components/schemas/ReactionCounts"
        totalReactionCount:
          type: integer
        currentUserReaction:
          type: string
          nullable: true
          enum: [like, heart, wow, angry]
        repostCount:
          type: integer
        commentCount:
          type: integer
        comments:
          type: array
          items:
            $ref: "#/components/schemas/PostComment"
        originalPost:
          allOf:
            - $ref: "#/components/schemas/Post"
          nullable: true
        author:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMedia"
      required: [id, userId, actorSurface, content, visibility, likeCount, isLiked, reactionCounts, totalReactionCount, repostCount, commentCount, comments, createdAt]
...
    ReactionCounts:
      type: object
      properties:
        like:
          type: integer
        heart:
          type: integer
        wow:
          type: integer
        angry:
          type: integer
      required: [like, heart, wow, angry]
...
    PostComment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        parentCommentId:
          type: integer
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/UserSummary"
          nullable: true
      required: [id, postId, userId, content, createdAt, updatedAt]
...
    PostMedia:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        type:
          type: string
        url:
          type: string
        title:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, postId, type, url, createdAt]
    PostsResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        limit:
          type: integer
        nextCursor:
          type: integer
          nullable: true
        hasMore:
          type: boolean
      required: [posts, total, limit, nextCursor, hasMore]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        isBookingInquiry:
          type: boolean
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        inquiry:
          $ref: "#/components/schemas/Inquiry"
          nullable: true
      required: [id, conversationId, senderId, content, isBookingInquiry, createdAt]
    SendMessageRequest:
      type: object
      properties:
        recipientId:
          type: integer
        content:
          type: string
      required: [recipientId, content]
    BookingInquiryRequest:
      type: object
      properties:
        eventDate:
          type: string
          nullable: true
        eventLocation:
          type: string
          nullable: true
        eventType:
          type: string
          nullable: true
        message:
          type: string
      required: [message]
    InquiryRequest:
      type: object
      properties:
        inquiryType:
          type: string
        eventDate:
          type: string
          nullable: true
        eventType:
          type: string
          nullable: true
        budget:
          type: string
          nullable: true
        projectDetails:
          type: string
          nullable: true
        externalUrl:
          type: string
          nullable: true
        message:
          type: string
      required: [message]
    Inquiry:
      type: object
      properties:
        inquiryType:
          type: string
        eventDate:
          type: string
          nullable: true
        eventType:
          type: string
          nullable: true
        budget:
          type: string
          nullable: true
        projectDetails:
          type: string
          nullable: true
        externalUrl:
          type: string
          nullable: true
...
    Group:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImageUrl:
          type: string
          nullable: true
        visibility:
          type: string
        category:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        owner:
          $ref: "#/components/schemas/UserSummary"
          nullable: true
        memberCount:
          type: integer
        postCount:
          type: integer
        isMember:
          type: boolean
        membersPreview:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
      required: [id, ownerId, name, slug, description, visibility, tags, createdAt, updatedAt, memberCount, postCount, isMember, membersPreview]
    GroupDetail:
      type: object
      properties:
        group:
          $ref: "#/components/schemas/Group"
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
      required: [group, posts]
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        coverImageUrl:
          type: string
          nullable: true
        visibility:
          type: string
        category:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required: [name, description]
...